refactor(ironfish): Move mint asset under wallet namespace
