+++ ironfish/src/rpc/clients/client.ts
@@ -62,7 +62,6 @@ import {
   UseAccountRequest,
   UseAccountResponse,
 } from '../routes'
-import { MintAssetRequest, MintAssetResponse } from '../routes/assets/mint'
 import {
   EstimateFeeRatesRequest,
   EstimateFeeRatesResponse,
@@ -85,6 +84,7 @@ import { GetRpcStatusRequest, GetRpcStatusResponse } from '../routes/rpc/getStat
 import { ExportAccountRequest, ExportAccountResponse } from '../routes/wallet/exportAccount'
 import { GetAccountStatusRequest, GetAccountStatusResponse } from '../routes/wallet/getStatus'
 import { ImportAccountRequest, ImportAccountResponse } from '../routes/wallet/importAccount'
+import { MintAssetRequest, MintAssetResponse } from '../routes/wallet/mintAsset'
 import { RemoveAccountRequest, RemoveAccountResponse } from '../routes/wallet/removeAccount'
 import { RescanAccountRequest, RescanAccountResponse } from '../routes/wallet/rescanAccount'
 
@@ -331,7 +331,10 @@ export abstract class RpcClient {
   }
 
   async mintAsset(params: MintAssetRequest): Promise<RpcResponseEnded<MintAssetResponse>> {
-    return this.request<MintAssetResponse>(`${ApiNamespace.asset}/mint`, params).waitForEnd()
+    return this.request<MintAssetResponse>(
+      `${ApiNamespace.wallet}/mintAsset`,
+      params,
+    ).waitForEnd()
   }
 
   async sendTransaction(
