+++ ironfish/src/rpc/routes/wallet/__fixtures__/mintAsset.test.ts.fixture
