+++ ironfish/src/rpc/routes/wallet/index.ts
@@ -12,9 +12,10 @@ export * from './getPublicKey'
 export * from './getStatus'
 export * from './getTransaction'
 export * from './getTransactions'
-export * from './sendTransaction'
 export * from './importAccount'
+export * from './mintAsset'
 export * from './removeAccount'
 export * from './rescanAccount'
+export * from './sendTransaction'
 export * from './sendTransaction'
 export * from './useAccount'
